refactor: split to new classes
