+++ packages/amplication-git-utils/src/git/github.service.ts
@@ -24,6 +24,8 @@ import {
 import { RemoteGitOrganization } from "./dto/remote-git-organization.dto";
 import { Branch } from "./dto/branch";
 import { EnumPullRequestMode } from "../types";
+import { BasicPullRequest } from "./github/BasicPullRequest";
+import { AccumulativePullRequest } from "./github/AccumulativePullRequest";
 
 const GITHUB_FILE_TYPE = "file";
 export const GITHUB_CLIENT_SECRET_VAR = "GITHUB_CLIENT_SECRET";
@@ -31,8 +33,6 @@ export const GITHUB_APP_APP_ID_VAR = "GITHUB_APP_APP_ID";
 export const GITHUB_APP_PRIVATE_KEY_VAR = "GITHUB_APP_PRIVATE_KEY";
 export const GITHUB_APP_INSTALLATION_URL_VAR = "GITHUB_APP_INSTALLATION_URL";
 export const UNEXPECTED_FILE_TYPE_OR_ENCODING = `Unexpected file type or encoding received`;
-
-const fileModeCode = "100644";
 
 type DirectoryItem = components["schemas"]["content-directory"][number];
 @Injectable()
@@ -201,15 +201,17 @@ export class GithubService {
     const amplicationIgnoreManger = new AmplicationIgnoreManger();
     await amplicationIgnoreManger.init(async (fileName) => {
       try {
-        return (
-          await this.getFile(
-            owner,
-            repo,
-            fileName,
-            baseBranchName,
-            installationId
-          )
-        ).content;
+        if (baseBranchName) {
+          return (
+            await this.getFile(
+              owner,
+              repo,
+              fileName,
+              baseBranchName,
+              installationId
+            )
+          ).content;
+        }
       } catch (error) {
         console.log("Repository does not have a .amplicationignore file");
         return "";
@@ -268,86 +270,32 @@ export class GithubService {
       })
     );
 
-    const branchInfo: {
-      repository: {
-        ref: {
-          associatedPullRequests: {
-            edges: [
-              {
-                node: {
-                  url: string;
-                  number: number;
-                };
-              }
-            ];
-          };
-        };
-      };
-    } = await octokit.graphql(
-      `
-      query ($owner: String!, $repo: String!, $head: String!) {
-        repository(name: $repo, owner: $owner) {
-          ref(qualifiedName: $head) {
-            associatedPullRequests(first: 1, states: OPEN) {
-              edges {
-                node {
-                  id
-                  number
-                  url
-                }
-              }
-            }
-          }
-        }
-      }`,
-      {
-        owner,
-        repo,
-        head,
-      }
-    );
-
-    const existingPullRequest =
-      branchInfo.repository.ref?.associatedPullRequests?.edges?.[0]?.node;
-
-    if (existingPullRequest) {
-      console.log("existingPR", existingPullRequest.url);
+    switch (mode) {
+      case EnumPullRequestMode.Accumulative:
+        return new AccumulativePullRequest().createPullRequest(
+          octokit,
+          owner,
+          repo,
+          prTitle,
+          prBody,
+          baseBranchName,
+          head,
+          files,
+          commitMessage
+        );
+      default:
+        return new BasicPullRequest().createPullRequest(
+          octokit,
+          owner,
+          repo,
+          prTitle,
+          prBody,
+          baseBranchName,
+          head,
+          files,
+          commitMessage
+        );
     }
-
-    if (!existingPullRequest) {
-      console.info("The PR does not exist, creating a new one");
-      // Returns a normal Octokit PR response
-      // See https://octokit.github.io/rest.js/#octokit-routes-pulls-create
-      const pr = await octokit.createPullRequest({
-        owner,
-        repo,
-        title: prTitle,
-        body: prBody,
-        base: baseBranchName /* optional: defaults to default branch */,
-        head,
-        update: true,
-        changes: [
-          {
-            /* optional: if `files` is not passed, an empty commit is created instead */
-            files: files,
-            commit: commitMessage,
-          },
-        ],
-      });
-      return pr.data.html_url;
-    }
-
-    console.info("The PR already exists, updating it");
-
-    await this.createCommit(
-      installationId,
-      owner,
-      repo,
-      commitMessage,
-      head,
-      files
-    );
-    return existingPullRequest.url;
   }
 
   private async getInstallationOctokit(
@@ -366,7 +314,7 @@ export class GithubService {
       url: repo.html_url,
       private: repo.private,
       fullName: repo.full_name,
-      admin: repo.permissions.admin,
+      admin: repo.permissions?.admin || false,
       defaultBranch: repo.default_branch,
     }));
   }
@@ -437,7 +385,7 @@ export class GithubService {
       full_name: fullName,
       permissions,
     } = response.data;
-    const { admin } = permissions;
+    const admin = permissions.admin || false;
     return {
       name,
       url: html_url,
@@ -470,113 +418,5 @@ export class GithubService {
       sha: refs.data.object.sha,
     });
     return { name: newBranchName, sha: branch.object.sha };
-  }
-
-  async isBranchExist(
-    installationId: string,
-    owner: string,
-    repo: string,
-    branch: string
-  ): Promise<boolean> {
-    try {
-      const refs = await this.getBranch(installationId, owner, repo, branch);
-      return Boolean(refs);
-    } catch (error) {
-      return false;
-    }
-  }
-
-  async getBranch(
-    installationId: string,
-    owner: string,
-    repo: string,
-    branch: string
-  ): Promise<Branch> {
-    const octokit = await this.getInstallationOctokit(installationId);
-    const refs = await octokit.rest.git.getRef({
-      owner,
-      repo,
-      ref: `heads/${branch}`,
-    });
-    return { sha: refs.data.object.sha, name: branch };
-  }
-
-  async createCommit(
-    installationId: string,
-    owner: string,
-    repo: string,
-    message: string,
-    branchName: string,
-    changes: any
-  ) {
-    const octokit = await this.getInstallationOctokit(installationId);
-    const changesArray = Object.entries(changes).map(([path, content]) => ({
-      path,
-      mode: fileModeCode,
-      content,
-    }));
-    const lastCommit = await this.getLastCommit(
-      installationId,
-      owner,
-      repo,
-      branchName
-    );
-
-    if (changesArray.length === 0) {
-      return;
-    }
-
-    const { data: tree } = await octokit.rest.git.createTree({
-      owner,
-      repo,
-      base_tree: lastCommit.commit.tree.sha,
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
-      tree: changesArray,
-    });
-
-    console.info(`Created tree for for ${owner}/${repo}`);
-
-    const { data: commit } = await octokit.rest.git.createCommit({
-      message,
-      owner,
-      repo,
-      tree: tree.sha,
-      parents: [lastCommit.sha],
-    });
-
-    console.info(`Created commit for ${owner}/${repo}`);
-
-    await octokit.rest.git.updateRef({
-      owner,
-      repo,
-      sha: commit.sha,
-      ref: `heads/${branchName}`,
-    });
-
-    console.info(`Updated branch ${branchName} for ${owner}/${repo}`);
-  }
-
-  async getLastCommit(
-    installationId: string,
-    owner: string,
-    repo: string,
-    branchName: string
-  ) {
-    const octokit = await this.getInstallationOctokit(installationId);
-    const branch = await this.getBranch(
-      installationId,
-      owner,
-      repo,
-      branchName
-    );
-    const [lastCommit] = (
-      await octokit.rest.repos.listCommits({
-        owner,
-        repo,
-        sha: branch.sha,
-      })
-    ).data;
-    return lastCommit;
   }
 }
